+++ src/helpers/game_mode.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct GameMode {
-  uint32_t map_index = 0;
-  uint32_t players_amount = 1;
-  uint32_t bots_amount = 0;
+  size_t map_index = 0;
+  size_t players_amount = 1;
+  size_t bots_amount = 0;
 };
